backup_accuracy_fixes_20250923_090400/requirements.txt
# Credit Risk Analysis - Production Requirements
# Azure and Databricks deployment ready

# Core Framework
flask>=3.0.0
flask-cors>=4.0.0
plotly>=5.15.0
pandas>=1.5.0
numpy>=1.24.0

# Data Processing  
openpyxl>=3.1.0
xlsxwriter>=3.1.0
rapidfuzz>=3.5.0

# Azure Dependencies
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-monitor-opentelemetry>=1.2.0

# Databricks dependencies
databricks-sdk>=0.12.0
pyspark>=3.4.0
delta-spark>=2.4.0

# Machine Learning
scikit-learn>=1.3.0
mlflow>=2.7.0

# LangGraph and AI
langgraph>=0.0.50
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# HTTP and API
requests>=2.31.0
httpx>=0.24.0

# File Operations and Locking
portalocker>=2.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
streamlit-aggrid>=1.1.8

# Future Phase 2 Dependencies (commented out for now)
# langchain>=0.1.0
# langgraph>=0.1.0
# openai>=1.0.0
# beautifulsoup4>=4.12.0
# requests>=2.31.0
# chromadb>=0.4.0

# Future Phase 3 Dependencies (commented out for now)
# databricks-sdk>=0.12.0
# mlflow>=2.8.0
# delta-spark>=3.0.0

# Legacy dependencies from earlier phases (keep for compatibility)
scikit-learn>=1.1.0
requests>=2.28.0
pydantic>=1.10.0
python-dotenv>=0.19.0
